--- CalculatorClone/CalculatorClone/Operator.swift
@@ -1,23 +1,0 @@
-//import SwiftUI
-//
-//func tappedButton(operator: currentOperator, button: CalculatorButtonContent) {
-//    switch button {
-//    case .divide, .multiply, .minus, .plus, .equal:
-//        if button == .divide {
-//            currentOperator = .divide
-//        } else if button == .multiply {
-//            currentOperator = .multiply
-//        } else if button == .minus {
-//            currentOperator = .minus
-//        } else if button == .plus {
-//            currentOperator = .plus
-//        } else if button == .equal {
-//
-//        }
-//    case .clear:
-//        value = "0"
-//    case .percent:
-//        value =
-//
-//    }
-//}
